+++ src/test/java/seedu/cakecollate/logic/commands/AddCommandTest.java
@@ -4,13 +4,9 @@ import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.cakecollate.logic.commands.CommandTestUtil.assertCommandFailure;
-import static seedu.cakecollate.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.cakecollate.testutil.Assert.assertThrows;
 import static seedu.cakecollate.testutil.TypicalIndexes.INDEX_FIRST_ORDER;
 import static seedu.cakecollate.testutil.TypicalIndexes.INDEX_THIRD_ORDER;
-import static seedu.cakecollate.testutil.TypicalOrderItems.getTypicalOrderItemsModel;
-import static seedu.cakecollate.testutil.TypicalOrders.getTypicalCakeCollate;
 
 import java.nio.file.Path;
 import java.util.ArrayList;
@@ -22,20 +18,16 @@ import org.junit.jupiter.api.Test;
 
 import javafx.collections.ObservableList;
 import seedu.cakecollate.commons.core.GuiSettings;
-import seedu.cakecollate.commons.core.Messages;
 import seedu.cakecollate.commons.core.index.IndexList;
 import seedu.cakecollate.logic.commands.exceptions.CommandException;
 import seedu.cakecollate.model.CakeCollate;
 import seedu.cakecollate.model.Model;
-import seedu.cakecollate.model.ModelManager;
 import seedu.cakecollate.model.ReadOnlyCakeCollate;
 import seedu.cakecollate.model.ReadOnlyUserPrefs;
-import seedu.cakecollate.model.UserPrefs;
 import seedu.cakecollate.model.order.Order;
 import seedu.cakecollate.model.orderitem.OrderItem;
 import seedu.cakecollate.testutil.AddOrderDescriptorBuilder;
 import seedu.cakecollate.testutil.OrderBuilder;
-import seedu.cakecollate.testutil.OrderItemBuilder;
 
 
 /**
@@ -47,7 +39,6 @@ import seedu.cakecollate.testutil.OrderItemBuilder;
  */
 
 public class AddCommandTest {
-    private Model model = new ModelManager(getTypicalCakeCollate(), getTypicalOrderItemsModel(), new UserPrefs());
     private IndexList nonNullIndexList = new IndexList(new ArrayList<>());
     // todo should i pass in empty index list instead of null? is there a good practice
 
@@ -61,53 +52,6 @@ public class AddCommandTest {
     public void constructor_nullAddOrderDescriptor_throwsNullPointerException() {
         assertThrows(NullPointerException.class, () -> new AddCommand(null, null));
         assertThrows(NullPointerException.class, () -> new AddCommand(nonNullIndexList, null));
-    }
-
-    @Test
-    public void invalidListIndex_uhh() throws CommandException {
-        // need to create own order item model -- or check edit tests
-        Model emptyModel = new ModelManager();
-        AddCommand.AddOrderDescriptor descriptor = new AddOrderDescriptorBuilder(new OrderBuilder().withOrderDescriptions().build()).build();
-
-        Command addCommand = new AddCommand(nonNullIndexList, descriptor);
-
-        assertCommandFailure(addCommand, emptyModel, Messages.MESSAGE_INVALID_ORDER_ITEM_INDEX);
-    }
-
-    @Test
-    public void execute_inputNewOrderDescription_addedToOrderItemModel() {
-        // need to create own order item model -- or check edit tests
-
-        String value = "cake that does not exist in model yet";
-        Order order = new OrderBuilder().withOrderDescriptions(value).build(); // cry names don't make sense
-        AddCommand.AddOrderDescriptor descriptor = new AddOrderDescriptorBuilder(order).build();
-
-        OrderItem newOrderItemToAdd = new OrderItemBuilder().withType(value).build();
-        // does this stuff go into dg, e.g. how to test - use builders
-
-        assert !model.hasOrderItem(newOrderItemToAdd)
-                : "corresponding order description shouldn't be in the order item model";
-
-        // int initialSize = model.getOrderItemLists.size();
-
-        Command addCommand = new AddCommand(null, descriptor);
-
-        Model expectedModel = new ModelManager(
-                new CakeCollate(model.getCakeCollate()),
-                //new OrderItems(model.getOrderItems()),
-                getTypicalOrderItemsModel(),
-                new UserPrefs()
-        );
-
-        expectedModel.addOrder(order);
-        expectedModel.addOrderItem(newOrderItemToAdd);
-        String expectedMessage = String.format(AddCommand.MESSAGE_SUCCESS, order);
-
-        assertCommandSuccess(addCommand, model, expectedMessage, expectedModel);
-        assertTrue(model.hasOrderItem(newOrderItemToAdd));
-        // assertTrue()
-        // assertEquals size has increased by one
-
     }
 
     @Test
@@ -164,6 +108,17 @@ public class AddCommandTest {
         // todo
 
     }
+
+
+
+    // ======== TESTS ADDED AFTER ORDER ITEM FEATURE ========
+
+
+
+    // ======== MODEL STUBS ========
+
+
+
 
     /**
      * A default model stub that have all of the methods failing.
